+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+1.0.7
+-----
+- Added empty billing and shipping information check in customer order controller
+- Fixed single product child push and delete task
+- Fixed checksum interface problems
+
 1.0.6
 -----
 - Added category cms title support
@@ -5,7 +11,7 @@
 - Added seo support for product image filenames
 - Added attribute for product active flag
 - Added customer group key mapping in foreign tables
-- Added exception for defaut EK customer group
+- Added exception for default EK customer group
 - Fixed active flag for product parents to 0
 - Fixed category and manufacturer image key mapping
 - Fixed key mapping missing table bug